functions/requirements.txt
flask
futsu
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
oauth2client
